requirement.txt
flask
opencv-python
pyserial
tqdm
pyserial-tools
